// api/model/workspaceDto.ts
/**
 * CPMS API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Amenity } from './amenity';
import { BankAccountInfo } from './bankAccountInfo';
import { DirectoryDto } from './directoryDto';
import { WorkspaceType } from './workspaceType';
import { WorkingHourDto } from './workingHourDto';
import { ContactInfo } from './contactInfo';
import { LeaseSharingInfo } from './leaseSharingInfo';
import { BillingInfo } from './billingInfo';
import { TariffDto } from './tariffDto';


export interface WorkspaceDto { 
    id?: string | null;
    name?: string | null;
    description?: string | null;
    workspaceType?: WorkspaceType;
    contactInfo?: ContactInfo;
    billingInfo?: BillingInfo;
    bankAccountInfo?: BankAccountInfo;
    extraNote?: string | null;
    imageUrl?: string | null;
    userId?: string | null;
    readonly chargePointsCount?: number;
    readableId?: string | null;
    directoryId?: string | null;
    directory?: DirectoryDto;
    directoryName?: string | null;
    directoryReadableId?: string | null;
    tariffId?: string | null;
    tariff?: TariffDto;
    leaseSharingInfo?: LeaseSharingInfo;
    workingHours?: Array<WorkingHourDto> | null;
    isReservationsEnabled?: boolean;
    amenities?: Array<Amenity> | null;
}
export namespace WorkspaceDto {
}