// api/model/roleDto.ts
/**
 * CPMS API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PolicyDto } from './policyDto';


export interface RoleDto { 
    id?: string | null;
    name?: string | null;
    description?: string | null;
    policies?: Array<PolicyDto> | null;
    parentRoleId?: string | null;
}